refactor: change folder directory for controlled-checkbox
